requirements.txt
chardet==3.0.4
Django==1.11.8
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==0.6.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.12
django-simple-captcha==0.5.6
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.10.3
idna==2.6
jdcal==1.3
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.4.9
Pillow==5.0.0
pycryptodome==3.5.1
pycryptodomex==3.6.0
PyYAML==3.12
requests==2.18.4
setuptools==38.4.0
six==1.11.0
South==1.0.2
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.22
wheel==0.30.0
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0